Descriptives: fix error outliers box plot with empty variable
Previously, an uninformative error was thrown if the
user asks to label outliers in the box plot while
one of the variables had no data. With this commit,
an empty plot is shown instead (this is the
customary behavior for all plots).

As I was not able to replicate this error in a test, no test is added in this commit.