linReg: make emmeans more robust for complex models
Previously, the reference grid that is used to compute the
estimated marginal means grew exponentially with the number
of predictors in the model. As a consequence, the reference
grid could grow so big that the emmeans analysis threw an error.

This commit marks all predictors that are not of interest for
the emmeans plot/table as so-called "nuisance parameters", so
that the reference grid averages over all these parameters and
thus remains small in size, even when the number of parameters
grow.

See https://cran.r-project.org/web/packages/emmeans/vignettes/messy-data.html