tTestIS: Allow rank biserial correlation to be negative
Rank biserial correlation previously used the given test-statistic
(the Mann-Whitney U) to calculate the rank biserial correlation.
However, because the Mann-Whitney U always takes the smallest of the
two calculated Us this means that the rank biserial correlation is
always positive and therefore doesn't indicate directionality of the
effect. This commit fixes this.