cfa: Fix exception handling bug
An update in lavaan has caused a lot of their error messages to
change. Because we catch a couple of exceptions based on these
messages, some of the error were not caught anymore. This also
cause one unit test to fail, flagging `jmv` for removal by
CRAN. To fix this, the message that we use to catch the error,
was changed so that it's compatible with both old and new versions
of lavaan.