rmAnova: improve singularity handling
When a singularity error occurs, there is now a main "warning message" that
describes what could be the problem and ways to mitigate this, and
for each row that is affected a short footnote is added.

Moreover, this commit contains a major refactor of the code to make handling
of all sphericity related results much easier.